+++ apps/e2e/forms-storybook/src/app/field-group/field-group.component.ts
@@ -26,7 +26,7 @@ export class FieldGroupComponent {
 
   constructor() {
     this.formGroup = this.#formBuilder.group({
-      name: new FormControl(undefined),
+      input: new FormControl(undefined),
     });
   }
 }