+++ libs/components/forms/src/lib/modules/file-attachment/file-drop.component.ts
@@ -156,6 +156,14 @@ export class SkyFileDropComponent implements OnDestroy {
   @Input({ transform: booleanAttribute })
   public labelHidden = false;
 
+  /**
+   * [Persistent inline help text](https://developer.blackbaud.com/skyux/design/guidelines/user-assistance#inline-help) that provides
+   * additional context to the user.
+   * @preview
+   */
+  @Input()
+  public hintText: string | undefined;
+
   @ViewChild('fileInput')
   public inputEl: ElementRef | undefined;