+++ libs/components/forms/src/lib/modules/shared/sky-forms-resources.module.ts
@@ -23,6 +23,7 @@ const RESOURCES: Record<string, SkyLibResources> = {
     skyux_character_count_over_limit: {
       message: 'You are over the character limit.',
     },
+    skyux_checkbox_group_required: { message: 'Required' },
     skyux_form_error_character_count: {
       message: 'Limit {0} to {1} character(s).',
     },
@@ -111,6 +112,7 @@ const RESOURCES: Record<string, SkyLibResources> = {
       message: '{0}: Please upload a file over {1}KB.',
     },
     skyux_file_attachment_label_no_file_chosen: { message: 'No file chosen.' },
+    skyux_file_attachment_required: { message: 'Required' },
     skyux_input_box_help_inline_aria_label: {
       message: 'Show help content for {0}',
     },